Remove multiple PSH/POP
I don't think anything else has this odd thing with multiple push/pop defined in the ISA that somehow works around ("all other specified registers being pushed or pulled") interrupt-causing instructions before triggering the interrupt and I quite dislike it. It smells like a nuisance for implementation and a set of sneaky rules for users. So! I'm getting rid of it. Begone, antifeature! (Just going to hope it is one and I don't figure out that it makes sense and regret this later...)